refactor(platform): move vmm_ipc switch under platform/cuda
use_vmm_api is CUDA-specific (cuMemCreate-backed KV pools), unlike
isolated_ipc, which is vendor-agnostic and stays at the platform root.
Hoist the is_use_vmm_api import in cuda/__init__.py to module level
alongside is_isolated_ipc since the module has no dependencies.

Refs #4805
