+++ tests/v1/test_vllm_mp_adapter.py
@@ -27,8 +27,8 @@ from lmcache.integration.vllm.vllm_multi_process_adapter import (
 )
 from lmcache.v1.multiprocess.group_view import EngineGroupInfo
 from lmcache.v1.multiprocess.protocol import RequestType
+from lmcache.v1.platform.cuda.vmm_ipc import is_use_vmm_api, set_use_vmm_api
 from lmcache.v1.platform.isolated_ipc import is_isolated_ipc, set_isolated_ipc
-from lmcache.v1.platform.vmm_ipc import is_use_vmm_api, set_use_vmm_api
 
 
 class FakeCudaEvent:
